Fixed bind to allow not specifying callback
When binding an exchange to a queue which already has messages in it, one might want to be able to not subscribe to the queue with default options, e.g. to be able to use 'ack'.
This fix enables you to only bind, so that you can subscribe later at any stage.